refactor signer
move all signer code into src/
