+++ test/testutils.ts
@@ -1,17 +1,12 @@
 import {ethers} from "hardhat";
 import {defaultAbiCoder, hexConcat, parseEther} from "ethers/lib/utils";
 import {Contract, ContractReceipt, Wallet} from "ethers";
-import {IERC20} from '../typechain'
+import {EntryPoint, EntryPoint__factory, IERC20, SimpleWallet__factory} from '../typechain'
 import {BytesLike} from "@ethersproject/bytes";
-import {
-  EntryPoint,
-  EntryPoint__factory,
-  SimpleWallet__factory
-} from "../typechain";
 import {expect} from "chai";
 import {Create2Factory} from "../src/Create2Factory";
+import {callDataCost, decodeRevertReason, rethrow} from "../src/userop/utils";
 
-export const AddressZero = ethers.constants.AddressZero
 export const HashZero = ethers.constants.HashZero
 export const ONE_ETH = parseEther('1');
 export const TWO_ETH = parseEther('2');
@@ -56,12 +51,6 @@ export function createWalletOwner(privkeyBase?: string): Wallet {
   // return new ethers.Wallet('0x'.padEnd(66, privkeyBase), ethers.provider);
 }
 
-export function callDataCost(data: string): number {
-  return ethers.utils.arrayify(data)
-    .map(x => x == 0 ? 4 : 16)
-    .reduce((sum, x) => sum + x)
-}
-
 export async function calcGasUsage(rcpt: ContractReceipt, entryPoint: EntryPoint, redeemerAddress?: string) {
   const actualGas = await rcpt.gasUsed
   const logs = await entryPoint.queryFilter(entryPoint.filters.UserOperationEvent(), rcpt.blockHash)
@@ -79,68 +68,6 @@ export async function calcGasUsage(rcpt: ContractReceipt, entryPoint: EntryPoint
 //helper function to create a constructor call to our wallet.
 export function WalletConstructor(entryPoint: string, owner: string): BytesLike {
   return new SimpleWallet__factory().getDeployTransaction(entryPoint, owner).data!
-}
-
-const panicCodes: { [key: string]: any } = {
-  //from https://docs.soliditylang.org/en/v0.8.0/control-structures.html
-  0x01: 'assert(false)',
-  0x11: 'arithmetic overflow/underflow',
-  0x12: 'divide by zero',
-  0x21: 'invalid enum value',
-  0x22: 'storage byte array that is incorrectly encoded',
-  0x31: '.pop() on an empty array.',
-  0x32: 'array sout-of-bounds or negative index',
-  0x41: 'memory overflow',
-  0x51: 'zero-initialized variable of internal function type'
-}
-
-//rethrow "cleaned up" exception.
-// - stack trace goes back to method (or catch) line, not inner provider
-// - attempt to parse revert data (needed for geth)
-// use with ".catch(rethrow())", so that current source file/line is meaningful.
-export function rethrow(): (e: Error) => void {
-  let callerStack = new Error().stack!.replace(/Error.*\n.*at.*\n/, '').replace(/.*at.* \(internal[\s\S]*/, '')
-
-  if (arguments[0]) {
-    throw new Error('must use .catch(rethrow()), and NOT .catch(rethrow)')
-  }
-  return function (e: Error) {
-    let solstack = e.stack!.match(/((?:.* at .*\.sol.*\n)+)/)
-    let stack = (solstack != null ? solstack[1] : '') + callerStack
-    // const regex = new RegExp('error=.*"data":"(.*?)"').compile()
-    const found = /error=.*?"data":"(.*?)"/.exec(e.message)
-    let message: string
-    if (found != null) {
-      const data = found![1]
-      message = decodeRevertReason(data) ?? e.message + ' - ' + data.slice(0, 100)
-    } else {
-      message = e.message
-    }
-    const err = new Error(message)
-    err.stack = 'Error: ' + message + '\n' + stack
-    throw err
-  }
-}
-
-export function decodeRevertReason(data: string, nullIfNoMatch = true): string | null {
-  const methodSig = data.slice(0, 10)
-  let dataParams = '0x' + data.slice(10);
-
-  if (methodSig == '0x08c379a0') {
-    const [err] = ethers.utils.defaultAbiCoder.decode(['string'], dataParams)
-    return `Error(${err})`
-  } else if (methodSig == '0x00fa072b') {
-    const [opindex, paymaster, msg] = ethers.utils.defaultAbiCoder.decode(['uint256', 'address', 'string'], dataParams)
-    return `FailedOp(${opindex}, ${paymaster != AddressZero ? paymaster : "none"}, ${msg})`
-  } else if (methodSig == '0x4e487b71') {
-    const [code] = ethers.utils.defaultAbiCoder.decode(['uint256'], dataParams)
-    return 'Panic(' + panicCodes[code] || code + ')'
-  }
-  if (!nullIfNoMatch) {
-    return data
-  }
-  return null
-
 }
 
 let currentNode: string = ''
